+++ packages/core/echo/echo-schema/src/ast/index.ts
@@ -5,4 +5,3 @@
 export * from './annotations';
 export * from './schema-validator';
 export * from './types';
-export * from './ref';
