+++ packages/core/echo/echo-schema/src/ast/ref.ts
@@ -6,10 +6,13 @@ import { type EncodedReference } from '@dxos/echo-protocol';
 import { S } from '@dxos/effect';
 import { DXN } from '@dxos/keys';
 
-import { type ObjectAnnotation, getObjectAnnotation, ReferenceAnnotationId, type JsonSchemaType } from '.';
-import { MutableSchema, StoredSchema } from '../mutable';
+import { MutableSchema } from '../mutable/mutable-schema';
+import { StoredSchema } from '../mutable/stored-schema';
 import { EXPANDO_TYPENAME } from '../object/expando';
-import { type WithId, type Ref, getTypename, type BaseObject } from '../types';
+import { type WithId, type Ref, type BaseObject } from '../types';
+import type { JsonSchemaType } from './types';
+import { getObjectAnnotation, ReferenceAnnotationId, type ObjectAnnotation } from './annotations';
+import { getTypename } from '../object/typename';
 
 /**
  * The `$id` field for an ECHO reference schema.
