--- packages/core/echo/echo-schema/src/mutable/index.ts
@@ -1,7 +1,0 @@
-//
-// Copyright 2024 DXOS.org
-//
-
-export * from './mutable-schema';
-export * from './runtime-schema-registry';
-export * from './types';
