+++ packages/core/echo/echo-schema/src/mutable/stored-schema.ts
@@ -1,0 +1,13 @@
+import { S } from '@dxos/effect';
+
+import { JsonSchemaType } from '../ast';
+import { TypedObject } from '../object/typed-object';
+/**
+ * Stored representation of a schema.
+ */
+// TODO(burdon): How to get the S.Schema object that this represents?
+export class StoredSchema extends TypedObject({ typename: 'dxos.org/type/Schema', version: '0.1.0' })({
+  typename: S.String,
+  version: S.String,
+  jsonSchema: JsonSchemaType,
+}) {}
