+++ packages/core/echo/echo-schema/src/object/index.ts
@@ -5,4 +5,5 @@
 export * from './typed-object';
 export * from './object-id';
 export * from './expando';
-export * from './meta'
+export * from './meta';
+export * from './typename';
