+++ packages/core/echo/echo-schema/src/types.ts
@@ -19,7 +19,6 @@ export const data = Symbol.for('@dxos/schema/Data');
 
 // TODO(burdon): Use consistently (with serialization utils).
 export const ECHO_ATTR_ID = '@id';
-export const ECHO_ATTR_TYPE = '@type';
 export const ECHO_ATTR_META = '@meta';
 
 //
@@ -126,33 +125,6 @@ export const requireTypeReference = (schema: S.Schema<any>): Reference => {
   }
 
   return typeReference;
-};
-
-/**
- * Querying the typename of the object.
- * The typename is the raw string without version: `example.com/type/Contact`.
- */
-// TODO(dmaretskyi): Convert to DXN.
-export const TYPENAME_SYMBOL = Symbol.for('@dxos/schema/Typename');
-
-/**
- * Sets the typename of the object without the version.
- */
-// TODO(dmaretskyi): Convert to DXN.
-export const getTypename = (obj: BaseObject): string | undefined => {
-  let typename = (obj as any)[TYPENAME_SYMBOL];
-
-  if (typename === undefined) {
-    typename = obj[ECHO_ATTR_TYPE];
-  }
-
-  if (typename === undefined) {
-    return undefined;
-  }
-  invariant(typeof typename === 'string');
-  invariant(!typename.startsWith('dxn:'));
-  invariant(!typename.includes('@'));
-  return typename;
 };
 
 // TODO(burdon): Move to client-protocol.
